perf: split landing page's API client from the admin/internal surface
client.ts (imported by LandingPage.tsx and Statistics.tsx) shipped the
full ~50-entry CONTRACT_ERROR_MESSAGES map and 15+ admin/blockchain/email
request builders, even though the landing page only ever calls
getStatistics and newsletterSubscribe.

Split into:
 - request.ts: shared retry/timeout/cache request layer
 - client.ts: the two endpoints the landing page actually uses
 - admin-client.ts: everything else (admin, blockchain, email, contract
   error messages), for whatever internal tooling needs that surface

A production build shows the two clients' code and string literals no
longer appear together; client.ts now type-checks to expose only
getStatistics and newsletterSubscribe, enforced by a test.

Note: a byte-for-byte comparison of the built .next/static/chunks output
against main showed no size difference for this app today, because
Turbopack's dead-code elimination already stripped the unused api
object's properties before this split. bundlewatch.config.json also
still targets classic webpack output paths (vendor.js, main*.js,
pages/_app*.js) that don't exist under this app's Turbopack/App Router
build, so it isn't actually wired up to catch a regression here either.
The split is still correct to make explicit and guards against future
regressions (e.g. a stray import pulling the admin surface back in), but
the bundlewatch config would need a separate fix to point at real
Turbopack chunk paths before it could verify this in CI.
